bug_fix

+++ dataset/custom_text.py
@@ -96,7 +96,7 @@ class CustomText(TextDataset):
         s_x = (max_x - min_x)  / poly_num
         s_y = (max_y - min_y) / poly_num
         for i in range(1, poly_num):
-            n_p = [min_x + int(s_x * i), min_y +int(s_y * i)]
+            n_p = [int(min_x + (s_x * i)), int(min_y +int(s_y * i))]
             
             nps.append(n_p)
         return nps
